style: apply safe ruff autofixes (unquote annotations, dict.fromkeys, ...)
23 mechanical fixes from the extended rule groups (UP037 unquoted annotations
now that __future__ annotations is imported, RUF022 sorted __all__, RUF046
redundant int() on len, SIM/C4 dict.fromkeys, UP035 collections.abc.Callable).
Behavior-preserving; no enforced-config change (kept as a one-off sweep).

+++ src/gradwave/postscf/stress_error.py
@@ -61,7 +61,7 @@ def _infer_kmesh(system) -> tuple[int, int, int]:
     strained system reproduces the run's k-point ordering.
     """
     kf = np.array([np.asarray(sph.k_frac, dtype=float) for sph in system.spheres])
-    return tuple(int(len(np.unique(np.round(kf[:, i] % 1.0, 6)))) for i in range(3))
+    return tuple(len(np.unique(np.round(kf[:, i] % 1.0, 6))) for i in range(3))
 
 
 @torch.no_grad()
